Problem Statement
Investigating Tacoma Narrows Bridge dynamics for safer designs through simulation using NumPy, Pandas, mathematical modeling, differential equations, simulation development and data analysis.
Developing a simulation model using four first-order equations to analyze bridge behavior and understand structural failure causes.
A mathematical model that attempts to capture the Tacoma Narrows Bridge incident was proposed by McKenna and Tuama [2001]. The goal is to explain how torsional, or twisting, oscillations can be magnified by forcing that is strictly vertical. Consider a roadway of width 2l hanging between two suspended cables, as in Figure 6.18(a). We will consider a two-dimensional slice of the bridge, ignoring the dimension of the bridge’s length for this model, since we are only interested in the side-to-side motion. At rest, the roadway hangs at a certain equilibrium height due to gravity; let y denote the current distance the center of the roadway hangs below this equilibrium.